veyron_ai_agency/requirements.txt
agency_swarm==0.3.0
openai==1.3.9
python-dotenv==1.0.0
gradio==5.20.1
requests==2.31.0
pillow>=10.0.0
pydantic>=2.0.0 